+++ lib/exception_notifier/squash_ruby/ruby.rb
